Рефакторинг адаптера альбомов
